[Refact]: Refactoring Controller and Service(#3)
