+++ src/main/java/com/click/auth/controller/AuthController.java
@@ -31,14 +31,6 @@ public class AuthController {
         return authService.findUserByCode(code);
     }
 
-    @GetMapping("/login/token")
-    public String getLoginToken(
-            @RequestParam("identity") String identity,
-            @RequestParam("type") UserIdentityType type
-    ){
-        return authService.generateLoginToken(identity, type);
-    }
-
     @PutMapping("/{id}")
     public void updateUser(
             @PathVariable("id") UUID id,
@@ -53,16 +45,4 @@ public class AuthController {
         authService.disableUser(id);
     }
 
-    @GetMapping("/token")
-    public String getUserToken(
-            @RequestParam("token") String token,
-            @RequestParam("password") String password
-    ){
-        return authService.generateUserToken(token, password);
-    }
-
-    @GetMapping("/token/{userToken}")
-    public UserTokenResponse getUserTokenInfo(@PathVariable("userToken") String userToken) {
-        return authService.parseUserToken(userToken);
-    }
 }
