+++ src/main/java/com/click/auth/service/AuthServiceImpl.java
@@ -3,13 +3,10 @@ package com.click.auth.service;
 import com.click.auth.domain.dto.response.LoginTokenResponse;
 import com.click.auth.domain.dto.request.UserCreateRequest;
 import com.click.auth.domain.dto.response.UserResponse;
-import com.click.auth.domain.dto.response.UserTokenResponse;
 import com.click.auth.domain.entity.User;
 import com.click.auth.domain.repository.UserRepository;
 import com.click.auth.domain.type.UserIdentityType;
-import com.click.auth.exception.LoginExpirationException;
 import com.click.auth.exception.NotFoundExcetion;
-import com.click.auth.exception.PasswordMatchException;
 import com.click.auth.util.FriendCodeUtils;
 import com.click.auth.util.JwtUtils;
 import jakarta.transaction.Transactional;
@@ -24,6 +21,18 @@ public class AuthServiceImpl implements AuthService{
     private final UserRepository userRepository;
     private final JwtUtils jwtUtils;
     private final FriendCodeUtils friendCodeUtils;
+
+    @Override
+    @Transactional
+    public String createUser(UserCreateRequest req) {
+        String code = friendCodeUtils.generateCode();
+        while (findUserByCode(code) != null){
+            code = friendCodeUtils.generateCode();
+        }
+        User user = req.toEntity(code);
+        userRepository.save(user);
+        return jwtUtils.createLoginToken(LoginTokenResponse.from(user));
+    }
 
     @Override
     public User findUserByIdentity(String identity, UserIdentityType type) {
@@ -54,41 +63,5 @@ public class AuthServiceImpl implements AuthService{
     public void disableUser(UUID id) {
         User user = findUserByUuid(id);
         user.disable();
-    }
-
-    @Override
-    @Transactional
-    public String createUser(UserCreateRequest req) {
-        String code = friendCodeUtils.generateCode();
-        while (findUserByCode(code) != null){
-            code = friendCodeUtils.generateCode();
-        }
-        User user = req.toEntity(code);
-        userRepository.save(user);
-        return jwtUtils.createLoginToken(LoginTokenResponse.from(user));
-    }
-
-    @Override
-    public String generateLoginToken(String identity, UserIdentityType type) {
-        User user = findUserByIdentity(identity, type);
-        return jwtUtils.createLoginToken(LoginTokenResponse.from(user));
-    }
-
-    @Override
-    public String generateUserToken(String accessToken, String password) {
-        LoginTokenResponse loginToken = jwtUtils.parseLoginToken(accessToken);
-        User user = userRepository.findById(loginToken.uuid()).orElseThrow(IllegalArgumentException::new);
-        if (!user.getUserPasswd().equals(password)) {
-            throw new PasswordMatchException();
-        }
-        if (!user.getUserTokenVersion().equals(loginToken.version())) {
-            throw new LoginExpirationException();
-        }
-        return jwtUtils.createUserToken(user);
-    }
-
-    @Override
-    public UserTokenResponse parseUserToken(String userToken) {
-        return jwtUtils.parseUserToken(userToken);
     }
 }
