--- src/main/java/com/click/auth/service/KakaoAuthService.java
@@ -1,7 +1,0 @@
-package com.click.auth.service;
-
-import com.click.auth.domain.dto.response.SocialLoginResponse;
-
-public interface KakaoAuthService {
-    SocialLoginResponse getUserToken(String AuthCode);
-}
