+++ src/main/java/com/click/auth/service/LoginService.java
@@ -1,0 +1,9 @@
+package com.click.auth.service;
+
+import com.click.auth.domain.dto.response.SocialLoginResponse;
+import com.click.auth.domain.type.UserIdentityType;
+
+public interface LoginService {
+    String generateLoginToken(String identity, UserIdentityType type);
+    SocialLoginResponse getUserTokenByKakao(String authCode);
+}
