+++ src/main/java/com/click/auth/service/TokenService.java
@@ -1,0 +1,8 @@
+package com.click.auth.service;
+
+import com.click.auth.domain.dto.response.UserTokenResponse;
+
+public interface TokenService {
+    String generateUserToken(String accessToken, String password);
+    UserTokenResponse parseUserToken(String userToken);
+}
